PhpRbac Conversion
Ok… so, now we're managing user access with [PhpRbac](http://phprbac.net/), which needs to be installed with [Composer](http://getcomposer.org). In this first pass, library.inc.php loads the RBAC (Role-based Access Control) object and checks to see if there is a user logged in to the session. If no user is logged in, a log is forced. Users can be managed via the contact page, and it's new subpane, the Database Roles page.

We are not yet doing a full check for permissions within pages, just blocking pages entirely for which a user doesn't have permission.

Signed-off-by: Seth Battis <seth@battis.net>